+++ themegen/lib/src/core/utils/where_util.dart
@@ -1,0 +1,3 @@
+extension WhereUtil on Never {
+  static bool whereNotSame<L, R>(L el, R el2) => el != el2;
+}