+++ themegen/lib/src/feature/ext/model/similar_ext.dart
@@ -1,8 +1,0 @@
-import 'package:analyzer/dart/element/type.dart';
-
-class SimilarExt {
-  final Set<DartType> types;
-  final int index;
-
-  SimilarExt(this.types, this.index);
-}