+++ themegen_annotation/CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Added the first version of the themegen_annotation package.
+- Added extensions